Improve modularity
+ minor renaming
+ minor tweaks
