+++ cnx.py
@@ -1,3 +1,6 @@
+"""Everything related to database connection"""
+
+
 import os
 from functools import wraps
