+++ starter_data/inventory_table.csv
