+++ static_data.py
@@ -1,17 +1,17 @@
 STARTER_DATA_FILES = {
-    "inventory": r"drafts/inventory_table.csv",
-    "locations": r"drafts/locations_table.csv",
-    "products": r"drafts/products_table.csv",
-    "products_to_suppliers": r"drafts/products_to_suppliers_table.csv",
-    "menu_items": r"drafts/menu_items_table.csv",
-    "proportions": r"drafts/proportions_table.csv",
-    "suppliers": r"drafts/suppliers_table.csv",
-    "contacts": r"drafts/contacts_table.csv",
-    "users": r"drafts/users_table.csv",
-    "purchase_orders": r"drafts/purchase_orders_table.csv",
-    "purchase_order_contents": r"drafts/purchase_order_contents_table.csv",
-    "access_levels": r"drafts/access_levels_table.csv",
-    "employees": r"drafts/employees_table.csv"
+    "inventory": r"starter_data/inventory_table.csv",
+    "locations": r"starter_data/locations_table.csv",
+    "products": r"starter_data/products_table.csv",
+    "products_to_suppliers": r"starter_data/products_to_suppliers_table.csv",
+    "menu_items": r"starter_data/menu_items_table.csv",
+    "proportions": r"starter_data/proportions_table.csv",
+    "suppliers": r"starter_data/suppliers_table.csv",
+    "contacts": r"starter_data/contacts_table.csv",
+    "users": r"starter_data/users_table.csv",
+    "purchase_orders": r"starter_data/purchase_orders_table.csv",
+    "purchase_order_contents": r"starter_data/purchase_order_contents_table.csv",
+    "access_levels": r"starter_data/access_levels_table.csv",
+    "employees": r"starter_data/employees_table.csv"
 }
